+++ src/float/approx_eq.rs
@@ -1,6 +1,6 @@
 use {super::epsilon::EPSILON, crate::approx_eq::ApproxEq};
 
-impl ApproxEq for f32 {
+impl ApproxEq for f64 {
     fn approx_eq(self, rhs: Self) -> bool {
         (self - rhs).abs() < EPSILON
     }
